Component for circuit page in both free and paid account + fixed some feedbacks from Pavlo
